Fix rasterization bugs related to 64-bit and to gcc optimizations.  The long pointers into 32-bit depth images needed to be an int pointer to be portable.  Current gcc optimizations break the int alias of the floating point grid column increments, which itself is an optimization that probably is not needed anymore, so revert to floating point math same as for rows.
git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@8403 3f9b6c58-9310-455a-aec6-8b30269ab333

+++ src/gp/RasterImage.c
@@ -108,7 +108,7 @@ ri_ClearBackground (DestImage *i)
 	for (ix = 0; ix < i->di_w; ix++)
 	{
 	    unsigned short *sdp = (unsigned short*) dp;
-	    unsigned long *ldp = (unsigned long*) dp;
+	    unsigned int *ldp = (unsigned int*) dp;
 		
 	    switch (i->di_bdepth)
 	    {
